# Intrinsicify SpanHelpers.IndexOfAny(char,...)