10% to 125% performance increase https://github.com/dotnet/runtime/pull/40729#issuecomment-673083373

Put them all in same PR as they are mostly identical

Intrinsicify IndexOfAny(char,char) https://github.com/dotnet/runtime/pull/40589
Intrinsicify IndexOfAny(char,char,char) https://github.com/dotnet/runtime/pull/40590
Intrinsicify IndexOfAny(char,char,char,char) https://github.com/dotnet/runtime/pull/40591
Intrinsicify IndexOfAny(char,char,char,char,char) https://github.com/dotnet/runtime/pull/40592

Resolves: #12094
Resolves: #12095
Resolves: #12096
Resolves: #12097

Resolves https://github.com/dotnet/runtime/issues/25023