@danmosemsft 2020-08-12
Thanks. Shall we close the others? I think it would be better.

You might consider running these microbenchmarks too
https://github.com/dotnet/performance/blob/8f00082e5f1ab8b86a98bf3bfc9c307a171912a3/src/benchmarks/micro/libraries/System.Memory/Span.cs#L72-L78