@benaadams 2020-08-12
+19%

```
| Faster                                                  | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------- | ---------:| ----------------:| ----------------:| ------- |
| BenchmarksGame.RegexRedux_1.RunBench                    |      1.11 |      45419250.00 |      40848366.67 |         |
| BenchmarksGame.RegexRedux_5.RunBench(options: Compiled) |      1.19 |       8419890.74 |       7057605.17 | bimodal |
| BenchmarksGame.RegexRedux_5.RunBench(options: None)     |      1.02 |      25805255.56 |      25379127.78 |         |
```

/cc @stephentoub