@benaadams 2020-08-12
+30% .. +65% for the span<char>, 512 chars
+75% for the string (pos 23, two items, 178 char string)

```
| Faster                                                    | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------- | ---------:| ----------------:| ----------------:| ------- |
| System.Tests.Perf_String.IndexOfAny                       |      1.75 |            14.00 |             7.98 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 512) |      1.65 |            24.25 |            14.71 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 512)   |      1.35 |            16.50 |            12.22 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 512)  |      1.30 |            29.68 |            22.83 |         |
```