@benaadams 2020-08-12
Smaller ones
```
| Slower                                                 | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 4) |      1.07 |             5.28 |             5.66 |         |

| Faster                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 8)    |      1.20 |             4.41 |             3.67 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 16)   |      1.44 |             5.66 |             3.94 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 24)   |      1.91 |             6.98 |             3.64 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 32)   |      2.25 |             8.83 |             3.92 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 48)   |      2.17 |             8.93 |             4.12 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 64)   |      1.92 |             9.33 |             4.85 |         |

| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 4)  |      1.18 |             4.07 |             3.44 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 16) |      1.63 |             7.23 |             4.44 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 24) |      2.12 |             8.66 |             4.08 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 32) |      2.11 |            10.21 |             4.84 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 48) |      2.31 |            10.76 |             4.65 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 64) |      2.01 |            11.33 |             5.64 |         |

| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 8)   |      1.11 |             6.70 |             6.06 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 16)  |      1.43 |             9.25 |             6.46 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 24)  |      1.84 |            11.55 |             6.28 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 48)  |      1.81 |            12.99 |             7.18 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 32)  |      1.77 |            12.44 |             7.04 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 64)  |      1.70 |            14.07 |             8.29 |         |
```