@danmosemsft 2020-08-12
Those numbers are great! I guess that is with AVX2. How do we test perf for CPU without that -- repeat with COMPlus_EnableAVX2=0 and COMPlus_EnableHWIntrinsic=0 (?)