@danmosemsft 2020-08-12
@tannergooding or @GrabYourPitchforks what configurations do you believe we should get perf numbers for? without AVX2 and without intrinsics?