@benaadams 2020-08-12
COMPlus_EnableAVX2=0
```
| Faster                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------:|
| BenchmarksGame.RegexRedux_1.RunBench                     |      1.13 |      45408050.00 |      40193533.33 |         |
| BenchmarksGame.RegexRedux_5.RunBench(options: Compiled)  |      1.16 |       7868740.74 |       6801046.97 |         |

| System.Tests.Perf_String.IndexOfAny                      |      1.43 |            14.66 |            10.22 |         |

| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 8)    |      1.40 |             4.38 |             3.12 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 16)   |      1.99 |             7.13 |             3.59 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 24)   |      1.48 |             7.69 |             5.18 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 32)   |      1.42 |             6.93 |             4.87 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 48)   |      1.50 |             8.52 |             5.69 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 64)   |      1.42 |             9.56 |             6.74 |         |

| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 4)  |      1.14 |             3.80 |             3.33 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 8)  |      1.16 |             4.78 |             4.13 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 16) |      1.87 |             8.54 |             4.57 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 24) |      1.89 |             9.14 |             4.85 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 32) |      1.63 |             9.26 |             5.67 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 48) |      1.53 |             9.93 |             6.48 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 64) |      1.42 |            10.68 |             7.53 |         |

| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 8)   |      1.10 |             6.37 |             5.80 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 16)  |      1.36 |             8.94 |             6.58 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 24)  |      1.67 |            10.86 |             6.51 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 32)  |      1.40 |            11.01 |             7.85 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 48)  |      1.16 |            10.26 |             8.82 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 64)  |      1.29 |            13.04 |            10.08 |         |
```