@benaadams 2020-08-12
`COMPlus_EnableSse2=0` + `COMPlus_EnableAVX2=0`
```
| Faster                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------:|
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 4)    |      1.18 |             2.65 |             2.25 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 8)    |      1.35 |             3.35 |             2.48 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 16)   |      1.26 |             4.75 |             3.77 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 24)   |      1.15 |             5.90 |             5.14 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 32)   |      1.13 |             7.15 |             6.34 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 48)   |      1.08 |             9.58 |             8.85 |         |
| System.Memory.Span<Char>.IndexOfAnyTwoValues(Size: 64)   |      1.06 |            12.01 |            11.35 |         |

| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 4)  |      1.20 |             3.03 |             2.52 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 8)  |      1.12 |             3.69 |             3.30 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 16) |      1.11 |             5.48 |             4.94 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 24) |      1.06 |             7.25 |             6.84 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 32) |      1.06 |             9.01 |             8.50 |         |
| System.Memory.Span<Char>.IndexOfAnyThreeValues(Size: 48) |      1.04 |            12.51 |            12.03 |         |

| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 16)  |      1.04 |             7.70 |             7.38 |         |
| System.Memory.Span<Char>.IndexOfAnyFourValues(Size: 32)  |      1.06 |            12.69 |            12.00 |         |
```